requirements.txt
streamlit>=1.32.0
python-pptx>=0.6.23
openpyxl>=3.1.2
pandas>=2.2.0
lxml>=5.0.0
python-dotenv>=1.0.0
